+++ src/Microsoft.AspNetCore.Http.Connections/Internal/ConnectionStatus.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+
+namespace Microsoft.AspNetCore.Http.Connections.Internal
+{
+    public enum ConnectionStatus
+    {
+        Inactive,
+        Active,
+        Disposed
+    }
+}
