+++ src/Microsoft.AspNetCore.Http.Connections/Internal/HttpConnectionContext.cs
@@ -8,14 +8,13 @@ using System.IO.Pipelines;
 using System.Security.Claims;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http.Features;
 using Microsoft.AspNetCore.Connections;
 using Microsoft.AspNetCore.Connections.Features;
 using Microsoft.AspNetCore.Http.Connections.Features;
-using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Http.Features;
 using Microsoft.Extensions.Logging;
 
-namespace Microsoft.AspNetCore.Http.Connections
+namespace Microsoft.AspNetCore.Http.Connections.Internal
 {
     public class HttpConnectionContext : ConnectionContext,
                                          IConnectionIdFeature,
@@ -280,13 +279,6 @@ namespace Microsoft.AspNetCore.Http.Connections
 
                 throw;
             }
-        }
-
-        public enum ConnectionStatus
-        {
-            Inactive,
-            Active,
-            Disposed
         }
 
         private static class Log
